+++ src/App.jsx
@@ -2,12 +2,12 @@
 
 import React, { useState, useEffect } from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import Login from "./screens/AuthScreen/Login";
-import Landing from "./screens/Landing";
+import Login from "./screens/LandingScreen/AuthScreen/Login";
+import Landing from "./screens/LandingScreen/Landing";
 import "./index.css";
 // import ".../../css/index.css";
 import ".../.././css/Sidebar.css";
-import Signup from "./screens/AuthScreen/Signup";
+import Signup from "./screens/LandingScreen/AuthScreen/Signup";
 import Dashboard from "./screens/DashboardScreen/Dashboard";
 import Sorters from "./screens/DashboardScreen/Sorters";
 import Customers from "./screens/DashboardScreen/Customers";
@@ -15,11 +15,11 @@ import Status from "./screens/DashboardScreen/Status";
 import Profile from "./screens/ProfileScreen/Profile";
 import SortingStatus from "./screens/DashboardScreen/SortingStatus";
 import Receipt from "./screens/ReceiptScreen/Receipt";
-import CompanyDetails from "./screens/AuthScreen/CompanyDetails";
-import ManageUsers from "./superadmin/users/ManageUsers";
+import CompanyDetails from "./screens/LandingScreen/AuthScreen/CompanyDetails";
+import ManageUsers from "./superadmin/ManageUsers";
 import api_endpoint from "./config";
-import About from "./screens/About";
-import ContactUs from "./screens/ContactUs";
+import About from "./screens/LandingScreen/About";
+import ContactUs from "./screens/LandingScreen/ContactUs";
 import CustomerArchived from "./screens/Archives/CustomerArchived";
 import StatusArchived from "./screens/Archives/StatusArchived";
 import Error from "./screens/error";
