+++ src/screens/LandingScreen/AuthScreen/CompanyDetails.jsx
@@ -2,11 +2,10 @@
 
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import Navbar from "../../component/Navbar";
 import { useForm } from "react-hook-form";
 import axios from "axios";
-import api_endpoint from "../../config";
-import Modal from "../../component/Modal";
+import api_endpoint from "../../../config";
+import Modal from "../../../component/Modal";
 
 const CompanyDetails = () => {
     const navigate = useNavigate();
