+++ src/screens/LandingScreen/AuthScreen/Login.jsx
@@ -2,12 +2,13 @@
 
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import Navbar from "../../component/Navbar";
 import { useForm } from "react-hook-form";
 import axios from "axios";
-import api_endpoint from "../../config";
-import ForgotPasswordModal from "../../component/ForgotPasswordModal";
+import api_endpoint from "../../../config";
+import Navbar from "../../../component/Navbar";
+import ForgotPasswordModal from "../../../component/ForgotPasswordModal";
 import Footer from "../Footer";
+
 
 const Login = () => {
   const navigate = useNavigate();
